requirements.txt
Django==4.0.6
psycopg2-binary==2.9.7
python-dotenv==1.1.1
whitenoise==6.9.0
djangorestframework==3.13.1
django-froala-editor==4.0.13
Pillow==9.2.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
python-decouple==3.8
python-dotenv
